requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic
